+++ crates/ui-components/src/lib.rs
@@ -38,6 +38,7 @@ pub mod clear_button;
 pub mod close_button;
 pub mod code;
 pub mod code_block;
+pub mod color_area;
 pub mod color_field;
 pub mod color_swatch;
 pub mod color_swatch_picker;
@@ -184,6 +185,7 @@ pub use clear_button::{ClearButton, ClearButtonVariant};
 pub use close_button::{CloseButton, CloseButtonSize, CloseButtonVariant};
 pub use code::{Code, CodeVariant};
 pub use code_block::{CodeBlock, CodeBlockMotion};
+pub use color_area::ColorArea;
 pub use color_field::ColorField;
 pub use color_swatch::{
     ColorSwatch, ColorSwatchAlpha, ColorSwatchRounding, ColorSwatchShape, ColorSwatchSize,